+++ test/extended_math_test.dart
@@ -1,8 +1,11 @@
+import 'dart:math';
+
 import 'package:extended_math/extended_math.dart';
 
 void main(List<String> args) {
-  final gen = Vector(<num>[4, 3]);
-  gen[1] = 8;
-  //gen.insert(1);
-  print(gen);
+  num equationFn(num value) {
+    return pow(value, 3) - 18 * value - 83;
+  }
+  final p = SecantMethod(equationFn, 2, 10, 0.001);
+  print(p.result());
 }